+++ packages/expo-router/src/getRoutesSSR.ts
@@ -17,7 +17,7 @@ export type Options = Omit<OptionsCore, 'getSystemRoute'>;
  */
 export function getRoutes(contextModule: RequireContext, options: Options = {}): RouteNode | null {
   return getRoutesCore(contextModule, {
-    getSystemRoute({ route, type, defaults }) {
+    getSystemRoute({ route, type, defaults, redirectConfig }) {
       if (route === '' && type === 'layout') {
         // Root layout when no layout is defined.
         return {
@@ -62,7 +62,7 @@ export function getRoutes(contextModule: RequireContext, options: Options = {}):
         return {
           ...defaults,
           loadRoute() {
-            return require('./getRoutesRedirects').getRedirectModule(route);
+            return require('./getRoutesRedirects').getRedirectModule(redirectConfig);
           },
         };
       }